Pregunta1/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-redis==5.4.0
djangorestframework==3.15.2
redis==5.2.1
sqlparse==0.5.3
